+++ src/main/ImGuiBackendDX9.cpp
@@ -37,17 +37,24 @@ extern "C" HINSTANCE ghInstance;
 
 // Implemented features:
 //  [X] Renderer: User texture binding. Use 'LPDIRECT3DTEXTURE9' as ImTextureID. Read the FAQ about ImTextureID!
-//  [X] Renderer: Multi-viewport support. Enable with 'io.ConfigFlags |= ImGuiConfigFlags_ViewportsEnable'.
-//  [X] Renderer: Support for large meshes (64k+ vertices) with 16-bit indices.
+//  [X] Renderer: Large meshes support (64k+ vertices) even with 16-bit indices (ImGuiBackendFlags_RendererHasVtxOffset).
+//  [X] Renderer: IMGUI_USE_BGRA_PACKED_COLOR support, as this is the optimal color encoding for DirectX9.
+//  [X] Renderer: Multi-viewport support (multiple windows). Enable with 'io.ConfigFlags |= ImGuiConfigFlags_ViewportsEnable'.
 
 // You can use unmodified imgui_impl_* files in your project. See examples/ folder for examples of using this.
 // Prefer including the entire imgui/ repository into your project (either as a copy or as a submodule), and only build the backends you need.
-// If you are new to Dear ImGui, read documentation from the docs/ folder + read the top of imgui.cpp.
-// Read online: https://github.com/ocornut/imgui/tree/master/docs
+// Learn about Dear ImGui:
+// - FAQ                  https://dearimgui.com/faq
+// - Getting Started      https://dearimgui.com/getting-started
+// - Documentation        https://dearimgui.com/docs (same as your local docs/ folder).
+// - Introduction, links and more at the top of imgui.cpp
 
 // CHANGELOG
 // (minor and older changes stripped away, please see git history for details)
-//  2021-XX-XX: Platform: Added support for multiple windows via the ImGuiPlatformIO interface.
+//  2024-XX-XX: Platform: Added support for multiple windows via the ImGuiPlatformIO interface.
+//  2024-10-07: DirectX9: Changed default texture sampler to Clamp instead of Repeat/Wrap.
+//  2024-02-12: DirectX9: Using RGBA format when supported by the driver to avoid CPU side conversion. (#6575)
+//  2022-10-11: Using 'nullptr' instead of 'NULL' as per our switch to C++11.
 //  2021-06-29: Reorganized backend to pull data from a single structure to facilitate usage with multiple-contexts (all g_XXXX access changed to bd->XXXX).
 //  2021-06-25: DirectX9: Explicitly disable texture state stages after >= 1.
 //  2021-05-19: DirectX9: Replaced direct access to ImDrawCmd::TextureId with a call to ImDrawCmd::GetTexID(). (will become a requirement)
@@ -74,6 +81,7 @@ struct ImGui_ImplDX9_Data
 	wil::com_ptr_nothrow<IDirect3DTexture9>        FontTexture;
 	int                                            VertexBufferSize = 5000;
 	int                                            IndexBufferSize = 10000;
+	bool                                           HasRgbaSupport;
 
 	ImGui_ImplDX9_Data()
 	{
@@ -103,8 +111,8 @@ static ImGui_ImplDX9_Data* ImGui_ImplDX9_GetBackendData()
 
 //----------------------------------------------------------------------------
 // Forward Declarations
-static void ImGui_ImplDX9_InitPlatformInterface();
-static void ImGui_ImplDX9_ShutdownPlatformInterface();
+static void ImGui_ImplDX9_InitMultiViewportSupport();
+static void ImGui_ImplDX9_ShutdownMultiViewportSupport();
 static void ImGui_ImplDX9_CreateDeviceObjectsForPlatformWindows();
 static void ImGui_ImplDX9_InvalidateDeviceObjectsForPlatformWindows();
 
@@ -120,41 +128,45 @@ void ImGui_ImplDX9_SetupRenderState(ImDrawData* drawData)
 	vp.Height = (DWORD)drawData->DisplaySize.y;
 	vp.MinZ = 0.0f;
 	vp.MaxZ = 1.0f;
-	bd->pd3dDevice->SetViewport(&vp);
+
+	auto device = bd->pd3dDevice;
+	device->SetViewport(&vp);
 
 	// Setup render state: fixed-pipeline, alpha-blending, no face culling, no depth testing, shade mode (for gradient)
-	bd->pd3dDevice->SetPixelShader(NULL);
-	bd->pd3dDevice->SetVertexShader(NULL);
-	bd->pd3dDevice->SetRenderState(D3DRS_FILLMODE, D3DFILL_SOLID);
-	bd->pd3dDevice->SetRenderState(D3DRS_SHADEMODE, D3DSHADE_GOURAUD);
-	bd->pd3dDevice->SetRenderState(D3DRS_ZWRITEENABLE, FALSE);
-	bd->pd3dDevice->SetRenderState(D3DRS_ALPHATESTENABLE, FALSE);
-	bd->pd3dDevice->SetRenderState(D3DRS_CULLMODE, D3DCULL_NONE);
-	bd->pd3dDevice->SetRenderState(D3DRS_ZENABLE, FALSE);
-	bd->pd3dDevice->SetRenderState(D3DRS_ALPHABLENDENABLE, TRUE);
-	bd->pd3dDevice->SetRenderState(D3DRS_BLENDOP, D3DBLENDOP_ADD);
-	bd->pd3dDevice->SetRenderState(D3DRS_SRCBLEND, D3DBLEND_SRCALPHA);
-	bd->pd3dDevice->SetRenderState(D3DRS_DESTBLEND, D3DBLEND_INVSRCALPHA);
-	bd->pd3dDevice->SetRenderState(D3DRS_SEPARATEALPHABLENDENABLE, TRUE);
-	bd->pd3dDevice->SetRenderState(D3DRS_SRCBLENDALPHA, D3DBLEND_ONE);
-	bd->pd3dDevice->SetRenderState(D3DRS_DESTBLENDALPHA, D3DBLEND_INVSRCALPHA);
-	bd->pd3dDevice->SetRenderState(D3DRS_SCISSORTESTENABLE, TRUE);
-	bd->pd3dDevice->SetRenderState(D3DRS_FOGENABLE, FALSE);
-	bd->pd3dDevice->SetRenderState(D3DRS_RANGEFOGENABLE, FALSE);
-	bd->pd3dDevice->SetRenderState(D3DRS_SPECULARENABLE, FALSE);
-	bd->pd3dDevice->SetRenderState(D3DRS_STENCILENABLE, FALSE);
-	bd->pd3dDevice->SetRenderState(D3DRS_CLIPPING, TRUE);
-	bd->pd3dDevice->SetRenderState(D3DRS_LIGHTING, FALSE);
-	bd->pd3dDevice->SetTextureStageState(0, D3DTSS_COLOROP, D3DTOP_MODULATE);
-	bd->pd3dDevice->SetTextureStageState(0, D3DTSS_COLORARG1, D3DTA_TEXTURE);
-	bd->pd3dDevice->SetTextureStageState(0, D3DTSS_COLORARG2, D3DTA_DIFFUSE);
-	bd->pd3dDevice->SetTextureStageState(0, D3DTSS_ALPHAOP, D3DTOP_MODULATE);
-	bd->pd3dDevice->SetTextureStageState(0, D3DTSS_ALPHAARG1, D3DTA_TEXTURE);
-	bd->pd3dDevice->SetTextureStageState(0, D3DTSS_ALPHAARG2, D3DTA_DIFFUSE);
-	bd->pd3dDevice->SetTextureStageState(1, D3DTSS_COLOROP, D3DTOP_DISABLE);
-	bd->pd3dDevice->SetTextureStageState(1, D3DTSS_ALPHAOP, D3DTOP_DISABLE);
-	bd->pd3dDevice->SetSamplerState(0, D3DSAMP_MINFILTER, D3DTEXF_LINEAR);
-	bd->pd3dDevice->SetSamplerState(0, D3DSAMP_MAGFILTER, D3DTEXF_LINEAR);
+	device->SetPixelShader(nullptr);
+	device->SetVertexShader(nullptr);
+	device->SetRenderState(D3DRS_FILLMODE, D3DFILL_SOLID);
+	device->SetRenderState(D3DRS_SHADEMODE, D3DSHADE_GOURAUD);
+	device->SetRenderState(D3DRS_ZWRITEENABLE, FALSE);
+	device->SetRenderState(D3DRS_ALPHATESTENABLE, FALSE);
+	device->SetRenderState(D3DRS_CULLMODE, D3DCULL_NONE);
+	device->SetRenderState(D3DRS_ZENABLE, FALSE);
+	device->SetRenderState(D3DRS_ALPHABLENDENABLE, TRUE);
+	device->SetRenderState(D3DRS_BLENDOP, D3DBLENDOP_ADD);
+	device->SetRenderState(D3DRS_SRCBLEND, D3DBLEND_SRCALPHA);
+	device->SetRenderState(D3DRS_DESTBLEND, D3DBLEND_INVSRCALPHA);
+	device->SetRenderState(D3DRS_SEPARATEALPHABLENDENABLE, TRUE);
+	device->SetRenderState(D3DRS_SRCBLENDALPHA, D3DBLEND_ONE);
+	device->SetRenderState(D3DRS_DESTBLENDALPHA, D3DBLEND_INVSRCALPHA);
+	device->SetRenderState(D3DRS_SCISSORTESTENABLE, TRUE);
+	device->SetRenderState(D3DRS_FOGENABLE, FALSE);
+	device->SetRenderState(D3DRS_RANGEFOGENABLE, FALSE);
+	device->SetRenderState(D3DRS_SPECULARENABLE, FALSE);
+	device->SetRenderState(D3DRS_STENCILENABLE, FALSE);
+	device->SetRenderState(D3DRS_CLIPPING, TRUE);
+	device->SetRenderState(D3DRS_LIGHTING, FALSE);
+	device->SetTextureStageState(0, D3DTSS_COLOROP, D3DTOP_MODULATE);
+	device->SetTextureStageState(0, D3DTSS_COLORARG1, D3DTA_TEXTURE);
+	device->SetTextureStageState(0, D3DTSS_COLORARG2, D3DTA_DIFFUSE);
+	device->SetTextureStageState(0, D3DTSS_ALPHAOP, D3DTOP_MODULATE);
+	device->SetTextureStageState(0, D3DTSS_ALPHAARG1, D3DTA_TEXTURE);
+	device->SetTextureStageState(0, D3DTSS_ALPHAARG2, D3DTA_DIFFUSE);
+	device->SetTextureStageState(1, D3DTSS_COLOROP, D3DTOP_DISABLE);
+	device->SetTextureStageState(1, D3DTSS_ALPHAOP, D3DTOP_DISABLE);
+	device->SetSamplerState(0, D3DSAMP_MINFILTER, D3DTEXF_LINEAR);
+	device->SetSamplerState(0, D3DSAMP_MAGFILTER, D3DTEXF_LINEAR);
+	device->SetSamplerState(0, D3DSAMP_ADDRESSU, D3DTADDRESS_CLAMP);
+	device->SetSamplerState(0, D3DSAMP_ADDRESSV, D3DTADDRESS_CLAMP);
 
 	// Setup orthographic projection matrix
 	// Our visible imgui space lies from draw_data->DisplayPos (top left) to draw_data->DisplayPos+data_data->DisplaySize (bottom right). DisplayPos is (0,0) for single viewport apps.
@@ -172,9 +184,9 @@ void ImGui_ImplDX9_SetupRenderState(ImDrawData* drawData)
 			0.0f,         0.0f,         0.5f,  0.0f,
 			(L+R)/(L-R),  (T+B)/(B-T),  0.5f,  1.0f
 		} } };
-		bd->pd3dDevice->SetTransform(D3DTS_WORLD, &mat_identity);
-		bd->pd3dDevice->SetTransform(D3DTS_VIEW, &mat_identity);
-		bd->pd3dDevice->SetTransform(D3DTS_PROJECTION, &mat_projection);
+		device->SetTransform(D3DTS_WORLD, &mat_identity);
+		device->SetTransform(D3DTS_VIEW, &mat_identity);
+		device->SetTransform(D3DTS_PROJECTION, &mat_projection);
 	}
 }
 
@@ -187,11 +199,13 @@ void ImGui_ImplDX9_RenderDrawData(ImDrawData* drawData)
 
 	// Create and grow buffers if needed
 	ImGui_ImplDX9_Data* bd = ImGui_ImplDX9_GetBackendData();
+	auto device = bd->pd3dDevice;
+
 	if (!bd->pVB || bd->VertexBufferSize < drawData->TotalVtxCount)
 	{
 		bd->VertexBufferSize = drawData->TotalVtxCount + 5000;
 
-		if (bd->pd3dDevice->CreateVertexBuffer(
+		if (device->CreateVertexBuffer(
 			bd->VertexBufferSize * sizeof(CUSTOMVERTEX),
 			D3DUSAGE_DYNAMIC | D3DUSAGE_WRITEONLY,
 			D3DFVF_CUSTOMVERTEX, D3DPOOL_DEFAULT, &bd->pVB, nullptr) < 0)
@@ -204,7 +218,7 @@ void ImGui_ImplDX9_RenderDrawData(ImDrawData* drawData)
 	{
 		bd->IndexBufferSize = drawData->TotalIdxCount + 10000;
 
-		if (bd->pd3dDevice->CreateIndexBuffer(
+		if (device->CreateIndexBuffer(
 			bd->IndexBufferSize * sizeof(ImDrawIdx),
 			D3DUSAGE_DYNAMIC | D3DUSAGE_WRITEONLY,
 			sizeof(ImDrawIdx) == 2 ? D3DFMT_INDEX16 : D3DFMT_INDEX32, D3DPOOL_DEFAULT, &bd->pIB, nullptr) < 0)
@@ -222,9 +236,9 @@ void ImGui_ImplDX9_RenderDrawData(ImDrawData* drawData)
 
 	// Backup the DX9 transform (DX9 documentation suggests that it is included in the StateBlock but it doesn't appear to)
 	D3DMATRIX lastWorld, lastView, lastProjection;
-	bd->pd3dDevice->GetTransform(D3DTS_WORLD, &lastWorld);
-	bd->pd3dDevice->GetTransform(D3DTS_VIEW, &lastView);
-	bd->pd3dDevice->GetTransform(D3DTS_PROJECTION, &lastProjection);
+	device->GetTransform(D3DTS_WORLD, &lastWorld);
+	device->GetTransform(D3DTS_VIEW, &lastView);
+	device->GetTransform(D3DTS_PROJECTION, &lastProjection);
 
 	// Allocate buffers
 	CUSTOMVERTEX* vtxDst;
@@ -244,10 +258,10 @@ void ImGui_ImplDX9_RenderDrawData(ImDrawData* drawData)
 	//  2) to avoid repacking vertices: #define IMGUI_OVERRIDE_DRAWVERT_STRUCT_LAYOUT struct ImDrawVert { ImVec2 pos; float z; ImU32 col; ImVec2 uv; }
 	for (int n = 0; n < drawData->CmdListsCount; n++)
 	{
-		const ImDrawList* cmdList = drawData->CmdLists[n];
-		const ImDrawVert* vtxSrc = cmdList->VtxBuffer.Data;
+		const ImDrawList* drawList = drawData->CmdLists[n];
+		const ImDrawVert* vtxSrc = drawList->VtxBuffer.Data;
 
-		for (int i = 0; i < cmdList->VtxBuffer.Size; i++)
+		for (int i = 0; i < drawList->VtxBuffer.Size; i++)
 		{
 			vtxDst->pos[0] = vtxSrc->pos.x;
 			vtxDst->pos[1] = vtxSrc->pos.y;
@@ -259,15 +273,15 @@ void ImGui_ImplDX9_RenderDrawData(ImDrawData* drawData)
 			vtxSrc++;
 		}
 
-		memcpy(idxDst, cmdList->IdxBuffer.Data, cmdList->IdxBuffer.Size * sizeof(ImDrawIdx));
-		idxDst += cmdList->IdxBuffer.Size;
+		memcpy(idxDst, drawList->IdxBuffer.Data, drawList->IdxBuffer.Size * sizeof(ImDrawIdx));
+		idxDst += drawList->IdxBuffer.Size;
 	}
 
 	bd->pVB->Unlock();
 	bd->pIB->Unlock();
-	bd->pd3dDevice->SetStreamSource(0, bd->pVB.get(), 0, sizeof(CUSTOMVERTEX));
-	bd->pd3dDevice->SetIndices(bd->pIB.get());
-	bd->pd3dDevice->SetFVF(D3DFVF_CUSTOMVERTEX);
+	device->SetStreamSource(0, bd->pVB.get(), 0, sizeof(CUSTOMVERTEX));
+	device->SetIndices(bd->pIB.get());
+	device->SetFVF(D3DFVF_CUSTOMVERTEX);
 
 	// Setup desired DX state
 	ImGui_ImplDX9_SetupRenderState(drawData);
@@ -279,10 +293,10 @@ void ImGui_ImplDX9_RenderDrawData(ImDrawData* drawData)
 	ImVec2 clipOff = drawData->DisplayPos;
 	for (int n = 0; n < drawData->CmdListsCount; n++)
 	{
-		const ImDrawList* cmdList = drawData->CmdLists[n];
-		for (int cmdIndex = 0; cmdIndex < cmdList->CmdBuffer.Size; cmdIndex++)
+		const ImDrawList* drawList = drawData->CmdLists[n];
+		for (int cmdIndex = 0; cmdIndex < drawList->CmdBuffer.Size; cmdIndex++)
 		{
-			const ImDrawCmd* pCmd = &cmdList->CmdBuffer[cmdIndex];
+			const ImDrawCmd* pCmd = &drawList->CmdBuffer[cmdIndex];
 			if (pCmd->UserCallback != nullptr)
 			{
 				// User callback, registered via ImDrawList::AddCallback()
@@ -290,7 +304,7 @@ void ImGui_ImplDX9_RenderDrawData(ImDrawData* drawData)
 				if (pCmd->UserCallback == ImDrawCallback_ResetRenderState)
 					ImGui_ImplDX9_SetupRenderState(drawData);
 				else
-					pCmd->UserCallback(cmdList, pCmd);
+					pCmd->UserCallback(drawList, pCmd);
 			}
 			else
 			{
@@ -300,17 +314,18 @@ void ImGui_ImplDX9_RenderDrawData(ImDrawData* drawData)
 				if (clipMax.x < clipMin.x || clipMax.y < clipMin.y)
 					continue;
 
-				// Apply Scissor/clipping rectangle, Bind texture, Draw
+				// Apply scissor/clipping rectangle
 				const RECT r = { (LONG)clipMin.x, (LONG)clipMin.y, (LONG)clipMax.x, (LONG)clipMax.y };
+				device->SetScissorRect(&r);
+
 				const LPDIRECT3DTEXTURE9 texture = (LPDIRECT3DTEXTURE9)pCmd->GetTexID();
-				bd->pd3dDevice->SetTexture(0, texture);
-				bd->pd3dDevice->SetScissorRect(&r);
-				bd->pd3dDevice->DrawIndexedPrimitive(D3DPT_TRIANGLELIST, pCmd->VtxOffset + globalVtxOffset, 0, (UINT)cmdList->VtxBuffer.Size, pCmd->IdxOffset + globalIdxOffset, pCmd->ElemCount / 3);
+				device->SetTexture(0, texture);
+				device->DrawIndexedPrimitive(D3DPT_TRIANGLELIST, pCmd->VtxOffset + globalVtxOffset, 0, (UINT)drawList->VtxBuffer.Size, pCmd->IdxOffset + globalIdxOffset, pCmd->ElemCount / 3);
 			}
 		}
 
-		globalIdxOffset += cmdList->IdxBuffer.Size;
-		globalVtxOffset += cmdList->VtxBuffer.Size;
+		globalIdxOffset += drawList->IdxBuffer.Size;
+		globalVtxOffset += drawList->VtxBuffer.Size;
 	}
 
 	// When using multi-viewports, it appears that there's an odd logic in DirectX9 which prevent subsequent windows
@@ -319,17 +334,33 @@ void ImGui_ImplDX9_RenderDrawData(ImDrawData* drawData)
 		bd->pd3dDevice->DrawIndexedPrimitive(D3DPT_TRIANGLELIST, 0, 0, 0, 0, 0);
 
 	// Restore the DX9 transform
-	bd->pd3dDevice->SetTransform(D3DTS_WORLD, &lastWorld);
-	bd->pd3dDevice->SetTransform(D3DTS_VIEW, &lastView);
-	bd->pd3dDevice->SetTransform(D3DTS_PROJECTION, &lastProjection);
+	device->SetTransform(D3DTS_WORLD, &lastWorld);
+	device->SetTransform(D3DTS_VIEW, &lastView);
+	device->SetTransform(D3DTS_PROJECTION, &lastProjection);
 
 	// Restore the DX9 state
 	d3d9StateBlock->Apply();
 }
 
+static bool ImGui_ImplDX9_CheckFormatSupport(LPDIRECT3DDEVICE9 pDevice, D3DFORMAT format)
+{
+	wil::com_ptr_nothrow<IDirect3D9> pd3d = nullptr;
+	if (pDevice->GetDirect3D(pd3d.addressof()) != D3D_OK)
+		return false;
+	D3DDEVICE_CREATION_PARAMETERS param = {};
+	D3DDISPLAYMODE mode = {};
+	if (pDevice->GetCreationParameters(&param) != D3D_OK || pDevice->GetDisplayMode(0, &mode) != D3D_OK)
+		return false;
+
+	// Font texture should support linear filter, color blend and write to render-target
+	bool support = (pd3d->CheckDeviceFormat(param.AdapterOrdinal, param.DeviceType, mode.Format, D3DUSAGE_DYNAMIC | D3DUSAGE_QUERY_FILTER | D3DUSAGE_QUERY_POSTPIXELSHADER_BLENDING, D3DRTYPE_TEXTURE, format)) == D3D_OK;
+	return support;
+}
+
 bool ImGui_ImplDX9_Init(IDirect3DDevice9* device)
 {
 	ImGuiIO& io = ImGui::GetIO();
+	IMGUI_CHECKVERSION();
 	IM_ASSERT(io.BackendRendererUserData == nullptr && "Already initialized a renderer backend!");
 
 	// Setup back-end capabilities flags
@@ -340,9 +371,10 @@ bool ImGui_ImplDX9_Init(IDirect3DDevice9* device)
 	io.BackendFlags |= ImGuiBackendFlags_RendererHasViewports;  // We can create multi-viewports on the Renderer side (optional)
 
 	bd->pd3dDevice = device;
+	bd->HasRgbaSupport = ImGui_ImplDX9_CheckFormatSupport(bd->pd3dDevice.get(), D3DFMT_A8B8G8R8);
 
 	if (io.ConfigFlags & ImGuiConfigFlags_ViewportsEnable)
-		ImGui_ImplDX9_InitPlatformInterface();
+		ImGui_ImplDX9_InitMultiViewportSupport();
 
 	return true;
 }
@@ -356,12 +388,34 @@ void ImGui_ImplDX9_Shutdown()
 	IM_ASSERT(bd != nullptr && "No renderer backend to shutdown, or already shutdown?");
 	ImGuiIO& io = ImGui::GetIO();
 
-	ImGui_ImplDX9_ShutdownPlatformInterface();
+	ImGui_ImplDX9_ShutdownMultiViewportSupport();
 	ImGui_ImplDX9_InvalidateDeviceObjects();
 
 	io.BackendRendererName = nullptr;
 	io.BackendRendererUserData = nullptr;
 	delete bd;
+}
+
+// Convert RGBA32 to BGRA32 (because RGBA32 is not well supported by DX9 devices)
+static void ImGui_ImplDX9_CopyTextureRegion(bool tex_use_colors, ImU32* src, int src_pitch, ImU32* dst, int dst_pitch, int w, int h)
+{
+#ifndef IMGUI_USE_BGRA_PACKED_COLOR
+	ImGui_ImplDX9_Data* bd = ImGui_ImplDX9_GetBackendData();
+	const bool convert_rgba_to_bgra = (!bd->HasRgbaSupport && tex_use_colors);
+#else
+	const bool convert_rgba_to_bgra = false;
+	IM_UNUSED(tex_use_colors);
+#endif
+	for (int y = 0; y < h; y++)
+	{
+		ImU32* src_p = (ImU32*)((unsigned char*)src + src_pitch * y);
+		ImU32* dst_p = (ImU32*)((unsigned char*)dst + dst_pitch * y);
+		if (convert_rgba_to_bgra)
+			for (int x = w; x > 0; x--, src_p++, dst_p++) // Convert copy
+				*dst_p = IMGUI_COL_TO_DX9_ARGB(*src_p);
+		else
+			memcpy(dst_p, src_p, w * 4); // Raw copy
+	}
 }
 
 bool ImGui_ImplDX9_CreateFontsTexture()
@@ -374,37 +428,19 @@ bool ImGui_ImplDX9_CreateFontsTexture()
 	int width, height, bytes_per_pixel;
 	io.Fonts->GetTexDataAsRGBA32(&pixels, &width, &height, &bytes_per_pixel);
 
-	// Convert RGBA32 to BGRA32 (because RGBA32 is not well supported by DX9 devices)
-#ifndef IMGUI_USE_BGRA_PACKED_COLOR
-	if (io.Fonts->TexPixelsUseColors)
-	{
-		ImU32* dstStart = (ImU32*)ImGui::MemAlloc((size_t)width * height * bytes_per_pixel);
-		for (ImU32* src = (ImU32*)pixels, *dst = dstStart, *dst_end = dstStart + (size_t)width * height; dst < dst_end; src++, dst++)
-			*dst = IMGUI_COL_TO_DX9_ARGB(*src);
-		pixels = (unsigned char*)dstStart;
-	}
-#endif
-
 	// Upload texture to graphics system
-	if (bd->pd3dDevice->CreateTexture(width, height, 1, D3DUSAGE_DYNAMIC, D3DFMT_A8R8G8B8, D3DPOOL_DEFAULT, &bd->FontTexture, nullptr) < 0)
+	if (bd->pd3dDevice->CreateTexture(width, height, 1, D3DUSAGE_DYNAMIC, bd->HasRgbaSupport ? D3DFMT_A8B8G8R8 : D3DFMT_A8R8G8B8, D3DPOOL_DEFAULT, &bd->FontTexture, nullptr) < 0)
 		return false;
 
 	D3DLOCKED_RECT tex_locked_rect;
 	if (bd->FontTexture->LockRect(0, &tex_locked_rect, nullptr, 0) != D3D_OK)
 		return false;
 
-	for (int y = 0; y < height; y++)
-		memcpy((unsigned char*)tex_locked_rect.pBits + (size_t)tex_locked_rect.Pitch * y, pixels + (size_t)width * bytes_per_pixel * y, (size_t)width * bytes_per_pixel);
+	ImGui_ImplDX9_CopyTextureRegion(io.Fonts->TexPixelsUseColors, (ImU32*)pixels, width * bytes_per_pixel, (ImU32*)tex_locked_rect.pBits, (int)tex_locked_rect.Pitch, width, height);
 	bd->FontTexture->UnlockRect(0);
 
 	// Store our identifier
 	io.Fonts->SetTexID((ImTextureID)bd->FontTexture.get());
-
-#ifndef IMGUI_USE_BGRA_PACKED_COLOR
-	if (io.Fonts->TexPixelsUseColors)
-		ImGui::MemFree(pixels);
-#endif
-
 	return true;
 }
 
@@ -438,7 +474,7 @@ void ImGui_ImplDX9_InvalidateDeviceObjects()
 void ImGui_ImplDX9_NewFrame()
 {
 	ImGui_ImplDX9_Data* bd = ImGui_ImplDX9_GetBackendData();
-	IM_ASSERT(bd != NULL && "Did you call ImGui_ImplDX9_Init()?");
+	IM_ASSERT(bd != nullptr && "Context or backend not initialized! Did you call ImGui_ImplDX9_Init()?");
 
 	if (!bd->FontTexture)
 		ImGui_ImplDX9_CreateDeviceObjects();
@@ -563,7 +599,7 @@ static void ImGui_ImplDX9_SwapBuffers(ImGuiViewport* viewport, void*)
 	}
 }
 
-static void ImGui_ImplDX9_InitPlatformInterface()
+static void ImGui_ImplDX9_InitMultiViewportSupport()
 {
 	ImGuiPlatformIO& platform_io = ImGui::GetPlatformIO();
 	platform_io.Renderer_CreateWindow = ImGui_ImplDX9_CreateWindow;
@@ -573,7 +609,7 @@ static void ImGui_ImplDX9_InitPlatformInterface()
 	platform_io.Renderer_SwapBuffers = ImGui_ImplDX9_SwapBuffers;
 }
 
-static void ImGui_ImplDX9_ShutdownPlatformInterface()
+static void ImGui_ImplDX9_ShutdownMultiViewportSupport()
 {
 	ImGui::DestroyPlatformWindows();
 }