+++ src/plugins/lua/bindings/lua_ImGuiUserTypes.cpp
@@ -263,7 +263,7 @@ void RegisterBindings_ImGuiUserTypes(sol::state_view lua)
 	imDrawList.set_function("AddNgonFilled", &ImDrawList::AddNgonFilled);
 	imDrawList.set_function("AddText", sol::overload(
 		[](ImDrawList& mThis, const ImVec2& pos, int col, std::string_view text) { mThis.AddText(pos, ImU32(col), text.data(), text.data() + text.size()); },
-		[](ImDrawList& mThis, const ImFont* font, float font_size, const ImVec2& pos, int col, std::string_view text) { mThis.AddText(font, font_size, pos, ImU32(col), text.data(), text.data() + text.size()); }));
+		[](ImDrawList& mThis, ImFont* font, float font_size, const ImVec2& pos, int col, std::string_view text) { mThis.AddText(font, font_size, pos, ImU32(col), text.data(), text.data() + text.size()); }));
 
 	imDrawList.set_function("AddPolyline", [](ImDrawList& mThis, const sol::as_table_t<std::vector<ImVec2>>& points, int col, int flags, float thickness)
 		{